Restruct lander code into separate file / terrain restruct fixes
