+++ source/main.lua
@@ -29,7 +29,7 @@ lovelyToasts = require("lib.lovelyToasts")
 -- gintScreenWidth = 1920
 -- gintScreenHeight = 1080
 
-
+Lander = require("lander")
 Terrain = require("terrain")
 
 gintScreenWidth = 1024-- 1920
@@ -76,446 +76,16 @@ gbolIsConnected = false			-- Will become true when received an acknowledgement f
 
 gbolDebug = true
 
-local function DoThrust(dt)
-
-	if garrLanders[1].fuel - dt >= 0 or (fun.LanderHasUpgrade(enum.moduleNamesThrusters) and garrLanders[1].fuel - (dt * 0.80) >= 0) then
-
-		garrLanders[1].engineOn = true
-		local angle_radian = math.rad(garrLanders[1].angle)
-		local force_x = math.cos(angle_radian) * dt
-		local force_y = math.sin(angle_radian) * dt
-		
-		-- adjust the thrust based on ship mass
-		local massratio = gintDefaultMass / fun.GetLanderMass()	-- less mass = higher ratio = more thrust = less fuel needed to move
-		if gbolDebug then garrMassRatio = massratio end			-- for debugging only
-		force_x = force_x * massratio
-		force_y = force_y * massratio
-
-		garrLanders[1].vx = garrLanders[1].vx + force_x
-		garrLanders[1].vy = garrLanders[1].vy + force_y
-
-		if fun.LanderHasUpgrade(enum.moduleNamesThrusters) then
-			garrLanders[1].fuel = garrLanders[1].fuel - (dt * 0.80)		-- efficient thrusters use 80% fuel compared to normal thrusters
-		else
-			garrLanders[1].fuel = garrLanders[1].fuel - (dt * 1)
-		end
-	else
-		-- no fuel to thrust
-		--! probably need to make a serious alert here
-	end
-end
-
-local function TurnLeft(dt)
--- rotate the lander anti-clockwise
-
-	garrLanders[1].angle = garrLanders[1].angle - (90 * dt)
-	if garrLanders[1].angle < 0 then garrLanders[1].angle = 360 end
-end
-
-local function TurnRight(dt)
--- rotate the lander clockwise
-
-	garrLanders[1].angle = garrLanders[1].angle + (90 * dt)
-	if garrLanders[1].angle > 360 then garrLanders[1].angle = 0 end
-
-end
-
-local function ThrustLeft(dt)
-
-	if fun.LanderHasUpgrade(enum.moduleNamesSideThrusters) then
-		local force_x = 0.5 * dt		--!
-		garrLanders[1].vx = garrLanders[1].vx - force_x
-		garrLanders[1].enginerighton = true						-- opposite engine is on
-		
-		garrLanders[1].fuel = garrLanders[1].fuel - force_x
-	end
-end
-
-local function ThrustRight(dt)
-
-	if fun.LanderHasUpgrade(enum.moduleNamesSideThrusters) then
-		local force_x = 0.5 * dt		--!
-		garrLanders[1].vx = garrLanders[1].vx + force_x
-		garrLanders[1].enginelefton = true						-- opposite engine is on
-		
-		garrLanders[1].fuel = garrLanders[1].fuel - force_x
-	end
-
-end
-
-local function MoveShip(Lander, dt)
-
-	Lander.x = Lander.x + Lander.vx
-	Lander.y = Lander.y + Lander.vy
-	
-	local leftedge = gintOriginX - (gintScreenWidth / 2)
-	if Lander.x < leftedge then Lander.x = leftedge end
-	
-	-- apply gravity
-	if Lander.landed == false then
-		Lander.vy = Lander.vy + (enum.constGravity * dt)
-	end
-	
-	if Lander.airborne then
-		gfltLandervy = Lander.vy		-- used to determine speed right before touchdown
-		gfltLandervx = Lander.vx
-	end
-	
-	-- capture a new smoke location every x seconds
-	gfltSmokeTimer = gfltSmokeTimer - dt
-	if gfltSmokeTimer <= 0 then
-		-- only produce smoke when not landed or any of the engines aren't firing
-		if (garrLanders[1].landed == false) and (garrLanders[1].engineOn or garrLanders[1].enginelefton or garrLanders[1].enginerighton) then
-			
-			gfltSmokeTimer = enum.constSmokeTimer	-- a new 'puff' is added when this timer expires (and above conditions are met)
-			
-			local mysmoke = {}
-			mysmoke.x = Lander.x
-			mysmoke.y = Lander.y
-			mysmoke.dt = 0			-- this timer will count up and determine which sprite to display
-			
-			table.insert(garrSmokeSprites, mysmoke)
-		end
-	end
-	
-end
-
-local function RefuelLander(objBase, dt)
--- drain fuel from the base and add it to the lander
--- objBase is an object/table item from garrObjects
-
-	local refuelamt = math.min(objBase.fuelqty, (garrLanders[1].fueltanksize - garrLanders[1].fuel), dt)
-
-	objBase.fuelqty = objBase.fuelqty - refuelamt
-	garrLanders[1].fuel = garrLanders[1].fuel + refuelamt
-	
-	-- disable the base if the tanks are empty
-	if objBase.fuelqty <= 0 then objBase.active = false end
-
-end
-
-local function PayLanderFromBase(objBase, fltDist)
--- pay some wealth based on distance to the base
--- objBase is an object/table item from garrObjects
--- fltDist is the distance from the base
-
-	local dist = math.abs(fltDist)
-	if objBase.paid == false then
-		garrLanders[1].wealth = cf.round(garrLanders[1].wealth + (100 - dist),0)
-		garrSound[4]:play()
-	end
-
-end
-
-local function PayLanderForControl(objBase)
-
-	if objBase.paid == false then
-		-- pay for a good vertical speed
-		garrLanders[1].wealth = cf.round(garrLanders[1].wealth + ((1 - gfltLandervy) * 100),0)
-		
-		-- pay for a good horizontal speed
-		garrLanders[1].wealth = cf.round(garrLanders[1].wealth + (0.60 - gfltLandervx * 100),0)
-		
-	end
-end
-
-local function CheckForDamage()
--- apply damage if vertical speed is too higher
-	
-	if garrLanders[1].vy > enum.constVYThreshold then
-		local excessspeed = garrLanders[1].vy - enum.constVYThreshold
-		garrLanders[1].health = garrLanders[1].health - (excessspeed * 100)
-	
-		if garrLanders[1].health < 0 then garrLanders[1].health = 0 end
-	end
 
 
-
-end
-
-local function CheckForContact(Lander,dt)
--- see if lander has contacted the ground
-
-	local LanderXValue = cf.round(Lander.x)
-	local groundYvalue
-	local onbase = fun.IsOnLandingPad(enum.basetypeFuel)
-
-	-- see if landed near a fuel base
-	-- bestdist could be a negative number meaning not yet past the base (but maybe really close to it)
-	local bestdist, bestbase = fun.GetDistanceToClosestBase(garrLanders[1].x, 2)		-- 2 = type of base = fuel
-	-- bestbase is an object/table item
-	-- add wealth based on alignment to centre of landing pad
-	if bestdist >= -80 and bestdist <= 40 then
-		onbase = true
-	end
-
-	-- get the height of the terrain under the lander
-	groundYvalue = cf.round(garrGround[LanderXValue],0)
-
-	-- check if lander is at or below the terrain
-	if Lander.y > groundYvalue - 8 then		-- the offset is the size of the lander image
-		Lander.landed = true
-
-		if onbase then
-			RefuelLander(bestbase,dt)
-			PayLanderFromBase(bestbase, bestdist)
-			
-			-- if lander was airborne then track that now it's not.
-			if Lander.airborne then
-				-- this is the first landing on this base so pay wealth based on vertical and horizontal speed
-				PayLanderForControl(bestbase)
-				bestbase.paid = true
-			end				
-		end
-		
-		if Lander.airborne then
-			-- a heavy landing will cause damage
-			CheckForDamage()
-			
-			Lander.airborne = false
-		end
-
-		Lander.vx = 0
-		if Lander.vy > 0 then Lander.vy = 0 end			
-		
-		-- check for game-over conditions
-		if garrLanders[1].fuel <= 1 and garrLanders[1].landed == true and onbase == false then
-			garrLanders[1].bolGameOver = true
-		end
-	else
-		Lander.landed = false
-		Lander.airborne = true
-	end
-end
-
-local function PlaySoundEffects()
-
-	if garrLanders[1].engineOn then
-		garrSound[1]:play()
-	else
-		garrSound[1]:stop()
-	end
-	
-	local fuelpercent = garrLanders[1].fuel / garrLanders[1].fueltanksize
-	
-	-- play alert if fuel is low (but not empty because that's just annoying)
-	if fuelpercent <= 0.33 and fuelpercent > 0.01 then		-- 1% because rounding (fuel is never actually zero)
-		garrSound[5]:play()
-	end
-end
-
-local function RecalcDefaultMass()
--- need to recalc the default mass
--- usually called after buying a module
-		local result = 0
-		-- all the masses are stored in this table so add them up
-		for i = 1, #garrLanders[1].mass do
-			result = result + garrLanders[1].mass[i]
-		end
-		return (result + garrLanders[1].fueltanksize)		-- mass of all the components + mass of fuel if the tank was full (i.e. fueltanksize)
-
-end
-
-local function PurchaseThrusters()
--- add fuel efficient thrusters to the lander
-
-	if garrLanders[1].wealth >= enum.moduleCostsThrusters then
-		for i = 1, #garrLanders[1].modules do
-			if garrLanders[1].modules[i] == enum.moduleNamesThrusters then
-				-- this module is already purchased. Abort
-				--! make a 'wrong' sound		
-				return
-			end
-		end
-		-- can purchase thrusters
-		
-		table.insert(garrLanders[1].modules, enum.moduleNamesThrusters)
-		garrLanders[1].wealth = garrLanders[1].wealth - enum.moduleCostsThrusters
-		
-		garrLanders[1].mass[1] = 115
-		
-		-- need to recalc the default mass
-		gintDefaultMass = RecalcDefaultMass()
-	else
-		-- play 'failed' sound
-		garrSound[6]:play()
-	end
-end
-
-local function PurchaseLargeTank()
--- add a larger tank to carry more fuelqty
-
-	if garrLanders[1].wealth >= enum.moduleCostsLargeTank then
-		for i = 1, #garrLanders[1].modules do
-			if garrLanders[1].modules[i] == enum.moduleNamesLargeTank then
-				-- this module is already purchased. Abort.
-				--! make a 'wrong' sound		
-				return
-			end
-		end
-		-- can purchase item
-		
-		table.insert(garrLanders[1].modules, enum.moduleNamesLargeTank)
-		garrLanders[1].wealth = garrLanders[1].wealth - enum.moduleCostsLargeTank
-		
-		garrLanders[1].fueltanksize = 32		-- an increase from the default (25)
-		garrLanders[1].mass[2] = 23
-		
-		-- need to recalc the default mass
-		gintDefaultMass = RecalcDefaultMass()
-	else
-		-- play 'failed' sound
-		garrSound[6]:play()		
-	end
-
-end
-
-local function PurchaseRangeFinder()
--- the rangefinder points to the nearest base
-
-	if garrLanders[1].wealth >= enum.moduleCostsRangeFinder then
-		for i = 1, #garrLanders[1].modules do
-			if garrLanders[1].modules[i] == enum.moduleNamesRangeFinder then
-				-- this module is already purchased. Abort.
-				--! make a 'wrong' sound		
-				return
-			end
-		end
-		-- can purchase item
-		
-		table.insert(garrLanders[1].modules, enum.moduleNamesRangeFinder)
-		garrLanders[1].wealth = garrLanders[1].wealth - enum.moduleCostsRangeFinder
-
-		garrLanders[1].mass[3] = 2	-- this is the mass of the rangefinder
-
-		-- need to recalc the default mass
-		gintDefaultMass = RecalcDefaultMass()		
-	else
-		-- play 'failed' sound
-		garrSound[6]:play()		
-	end
-
-end
-
-local function PurchaseSideThrusters()
-
-	if garrLanders[1].wealth >= enum.moduleCostSideThrusters then
-		if not fun.LanderHasUpgrade(enum.moduleNamesSideThrusters) then
-			table.insert(garrLanders[1].modules, enum.moduleNamesSideThrusters)
-			garrLanders[1].wealth = garrLanders[1].wealth - enum.moduleCostSideThrusters
-
-			garrLanders[1].mass[4] = enum.moduleMassSideThrusters	-- this is the mass of the side thrusters
-
-			-- need to recalc the default mass
-			gintDefaultMass = RecalcDefaultMass()	
-		end
-	else
-		-- play 'failed' sound
-		garrSound[6]:play()		
-	end
-end
-
-local function HandleSockets()
-	
-	-- add lander info to the outgoing queue
-	local msg = {}
-	msg.x = garrLanders[1].x
-	msg.y = garrLanders[1].y
-	msg.angle = garrLanders[1].angle
-	msg.name = garrLanders[1].name
-	-- ** msg is set here and sent below
-	
-	if gbolIsAHost then
-		ss.HostListenPort()
-		
-		-- get just one item from the queue and process it
-		repeat
-			local incoming = ss.GetItemInHostQueue()		-- could be nil
-			if incoming ~= nil then
-				if incoming.name == "ConnectionRequest" then
-					gbolIsConnected = true
-					msg = {}
-					msg.name = "ConnectionAccepted"
-
-				else
-					garrLanders[2] = {}			--! super big flaw: this hardcodes garrLanders[2]. 
-					garrLanders[2].x = incoming.x
-					garrLanders[2].y = incoming.y
-					garrLanders[2].angle = incoming.angle
-					garrLanders[2].name = incoming.name
-				end	
-			end
-		until incoming == nil
-			
-		ss.AddItemToHostOutgoingQueue(msg)
-		ss.SendToClients()
-		msg = {}
-	end
-	
-	if gbolIsAClient then
-		ss.ClientListenPort()
-		
-		-- get item from the queue and process it
-		repeat
-			local incoming = ss.GetItemInClientQueue()		-- could be nil
-			if incoming ~= nil then
-				if incoming.name == "ConnectionAccepted" then
-					gbolIsConnected = true
-					if garrCurrentScreen[#garrCurrentScreen] == "MainMenu" then
-						fun.SaveGameSettings()
-						fun.AddScreen("World")
-					end
-				else	
-					garrLanders[2] = {}
-					garrLanders[2].x = incoming.x
-					garrLanders[2].y = incoming.y
-					garrLanders[2].angle = incoming.angle
-					garrLanders[2].name = incoming.name
-				end
-			end
-		until incoming == nil
-
-		ss.AddItemToClientOutgoingQueue(msg)	-- Lander[1]
-		ss.SendToHost()
-		msg = {}
-	end
-		
-end
-
-local function UpdateSmoke(dt)
--- each entry in the smoke table tracks it's own life (dt) so it knows when to expire
-
-	for k,v in pairs(garrSmokeSprites) do
-		v.dt = v.dt + (dt * 6)	-- 6 seems to give a good effect
-		if v.dt > 8 then		-- the sprite sheet has 8 images
-			table.remove(garrSmokeSprites,k)
-		end
-	end
-end
-
-function love.keypressed( key, scancode, isrepeat)
+function love.keypressed(key, scancode, isrepeat)
 --! don't like how there are keypressed events here as well as in love.update
 	if key == "escape" then
 		fun.RemoveScreen()
 	end
-	
-	if fun.IsOnLandingPad(2) then	-- 2 = base type (fuel)
-		if key == "1" then			 
-			PurchaseThrusters()
-		end
-	
-		if key == "2" then			
-			PurchaseLargeTank()
-		end	
-		
-		if key == "3" then			
-			PurchaseRangeFinder()
-		end			
-		if key == "4" then			
-			PurchaseSideThrusters()
-		end		
-	end
-	
+
+	Lander.keypressed(key, scancode, isrepeat)
+
 	if key == "r" then
 		if garrLanders[1].bolGameOver then
 			fun.ResetGame()
@@ -542,7 +112,7 @@ function love.load()
 	fun.ResetGame()
 	
 	-- capture the 'normal' mass of the lander into a global variable
-	gintDefaultMass = fun.GetLanderMass()
+	gintDefaultMass = Lander.getMass()
 	
 	-- stills/images
 	--! should make these numbers enums one day
@@ -629,46 +199,17 @@ function love.update(dt)
 	local strCurrentScreen = garrCurrentScreen[#garrCurrentScreen]
 	
 	if strCurrentScreen == "MainMenu" or strCurrentScreen == "Credits" or strCurrentScreen == "Settings" then
-		HandleSockets()
+		Lander.handleSockets()
 		Slab.Update(dt)		
 	end
 	
 	if strCurrentScreen == "World" then
 
-		if love.keyboard.isDown("up") or love.keyboard.isDown("w") or love.keyboard.isDown("kp8") then
-			DoThrust(dt)
-		end
-		if love.keyboard.isDown("left") or love.keyboard.isDown("a") or love.keyboard.isDown("kp4") then
-			TurnLeft(dt)
-		end
-		if love.keyboard.isDown("right") or love.keyboard.isDown("d") or love.keyboard.isDown("kp6") then
-			TurnRight(dt)
-		end
-		if love.keyboard.isDown("q") or love.keyboard.isDown("kp7") then
-			ThrustLeft(dt)
-		end
-		if love.keyboard.isDown("e") or love.keyboard.isDown("kp9") then
-			ThrustRight(dt)
-		end		
-		if love.keyboard.isDown("p") then
-			fun.AddScreen("Pause")
-		end
-		if love.keyboard.isDown("o") then
-			fun.AddScreen("Settings")
-		end
-		
-		
-		MoveShip(garrLanders[1], dt)		--! some really inconsistent use of parameters here
-		
-		UpdateSmoke(dt)
-		
-		PlaySoundEffects()
-		
-		CheckForContact(garrLanders[1], dt)
+		Lander.update(dt)
 		
 		gLandingLightsAnimation:update(dt)
 		
-		HandleSockets()
+		Lander.handleSockets()
 	end
 	
 	lovelyToasts.update(dt)		-- can potentially move this with the Slab.Update as it is only used on the main menu
